+++ src/components/Confirm/Confirm.tsx
@@ -1,16 +1,15 @@
 import React, { FC } from 'react'
-import { CustomizationProvider } from '../../features/CustomizationProvider/CustomizationProvider'
-import { ThemeProvider } from '../../features/ThemeProvider/ThemeProvider'
 import { Button } from './../Button/Button'
 import { Modal } from '../Modal/Modal'
 import './Confirm.css'
 
-interface ConfirmProps extends CustomizationProvider {
+interface ConfirmProps {
   title: string
   isShown: boolean
   onConfirm: () => void
   onCancel?: () => void
   onClose: () => void
+  className?: string
 }
 
 export const Confirm: FC<ConfirmProps> = ({
@@ -32,20 +31,18 @@ export const Confirm: FC<ConfirmProps> = ({
   }
 
   return (
-    <ThemeProvider>
-      <Modal 
-        isShown={isShown} 
-        title={title} 
-        onClose={onClose} 
-        className={className}
-      >
-        <div className="__Buttons__">
-          <Button onClick={cancelHandler} outlined>
-            Cancel
-          </Button>
-          <Button onClick={confirmHandler}>Confirm</Button>
-        </div>
-      </Modal>
-    </ThemeProvider>
+    <Modal 
+      isShown={isShown} 
+      title={title} 
+      onClose={onClose} 
+      className={className}
+    >
+      <div className="__Buttons__">
+        <Button onClick={cancelHandler} outlined>
+          Cancel
+        </Button>
+        <Button onClick={confirmHandler}>Confirm</Button>
+      </div>
+    </Modal>
   )
 }