+++ src/components/Date/const/defaultTimeZone.ts
@@ -1,0 +1,3 @@
+import { Region } from "../types/Region"
+
+export const defaultTimeZone: Region = 'Europe'