+++ src/components/Date/utils/fillDateTemplate.ts
@@ -1,13 +1,12 @@
-import { Region } from './Region.type'
+import { defaultTimeZone } from '../const/defaultTimeZone'
+import { timeFormatMap } from '../const/timeFormatMap'
 import { parseTime } from './parseTime'
-import { timeFormatMap } from './timeFormatMap'
+import { Region } from '../types/Region'
 
 export function fillDateTemplate(time: number, region: Region) {
   const { day, month, year } = parseTime(time)
 
-  if (!timeFormatMap[region]) region = 'Europe'
-
-  return timeFormatMap[region]
+  return timeFormatMap[region || defaultTimeZone]
     .replace(/DD/, day)
     .replace(/MM/, month)
     .replace(/YY/, year)