+++ src/components/Modal/Modal.tsx
@@ -1,15 +1,14 @@
 import React, { FC, ReactNode } from 'react'
-import { CustomizationProvider } from '../../features/CustomizationProvider/CustomizationProvider'
 import { BackgroundLayer } from './../BackgroundLayer/BackgroundLayer'
-import { ThemeProvider } from '../../features/ThemeProvider/ThemeProvider'
 import cn from 'classnames'
 import './Modal.css'
 
-interface ModalProps extends CustomizationProvider {
+interface ModalProps {
   title: string
   children?: ReactNode
   isShown: boolean
   onClose: () => void
+  className?: string
 }
 
 export const Modal: FC<ModalProps> = ({
@@ -19,19 +18,21 @@ export const Modal: FC<ModalProps> = ({
   onClose,
   className,
 }) => {
-  const classNames = cn('__Modal__', !isShown && '__Modal-Hidden__', className)
+  const classNames = cn(
+    '__Modal__', 
+    !isShown && '__Modal-Hidden__', 
+    className
+  )
 
   return (
-    <ThemeProvider>
-      <BackgroundLayer 
-        onClick={onClose} 
-        isShown={isShown}
-      >
-        <div className={classNames}>
-          <div className="__Title__">{title}</div>
-          {children}
-        </div>
-      </BackgroundLayer>
-    </ThemeProvider>
+    <BackgroundLayer 
+      onClick={onClose} 
+      isShown={isShown}
+    >
+      <div className={classNames}>
+        <div className="__Title__">{title}</div>
+        {children}
+      </div>
+    </BackgroundLayer>
   )
 }