+++ src/components/Title/Title.tsx
@@ -1,15 +1,19 @@
 import React, { FC, ReactNode } from 'react'
-import { CustomizationProvider } from '../../features/CustomizationProvider/CustomizationProvider'
 import cn from 'classnames'
 import './Title.css'
 
-interface TitleProps extends CustomizationProvider {
+interface TitleProps {
   children: ReactNode
   bold?: boolean
+  className?: string
 }
 
 export const Title: FC<TitleProps> = ({ children, bold, className }) => {
-  const classNames = cn('__Title__', bold && '__Bold__', className)
+  const classNames = cn(
+    '__Title__', 
+    bold && '__Bold__', 
+    className
+  )
 
   return (
     <div className={classNames}>